# Conflicts:
#	src/main/java/com/pse/testserver/entities/Category.java
#	src/main/java/com/pse/testserver/entities/Comment.java
#	src/main/java/com/pse/testserver/entities/Event.java
#	src/main/java/com/pse/testserver/entities/Group.java
#	src/main/java/com/pse/testserver/entities/Post.java
#	src/main/java/com/pse/testserver/entities/Subcategory.java
#	src/main/java/com/pse/testserver/entities/User.java
#	src/main/resources/application.properties

+++ src/main/java/com/pse/testserver/repository/EventRepository.java
@@ -20,7 +20,7 @@ public interface EventRepository extends CrudRepository<Event, Integer> {
      * @param id unique id of the searched event
      * @return event with the defined id
      */
-    Event findById(long id);
+    Event findById(int id);
 
     /**
      * Find all events from the database by date.
